- w.End inclusive (corresponding point still excluded) - interval: int64->float64 - col/row: int64->int

+++ bow/aggregation/arithmeticmean.go
@@ -12,7 +12,7 @@ func ArithmeticMean(col string) bow.ColumnAggregation {
 			}
 
 			var sum float64
-			for i := int64(0); i < w.Bow.NumRows(); i++ {
+			for i := 0; i < w.Bow.NumRows(); i++ {
 				value, ok := w.Bow.GetFloat64(col, int(i))
 				if !ok {
 					continue
